fix: getting linkedin logo
